move IssuerSearchComponent into _shared module
