+++ src/app/report/report-parameter/_shared/issuer-search/issuer-search.component.scss
